Refactor TV utils, fix clearNativeFolder.ts
